feat(devtools): move vfs handler to devtools
