+++ framework/ios/base/bridge/HippyBridge.h
@@ -238,8 +238,6 @@ HP_EXTERN NSString *HippyBridgeModuleNameForClass(Class bridgeModuleClass);
 
 @property (nonatomic, assign) BOOL debugMode;
 
-@property (nonatomic, assign) NSURL *debugURL;
-
 @property (nonatomic, strong) NSString *appVerson;  //
 
 @property (nonatomic, assign) HPInvalidateReason invalidateReason;
